Add some initial wikidata/wikibase views
Q5, wikibase:item
P18, P19, P21, P26, P2048

The underlying model is quite difficult to implement within the existing
view structure, which leads to my questioning of whether the views should
in fact adapted/re-used.

A better route might be that add yet another layer with basic components
which are filled in by the above layers, so logic can be shared (albeit
at the cost of mounting another component).

I do like their model about describing statements, effectively making
their data more truthful. But it is really annoying that the properties
on the `Statement` model aren't isolationally deterministic (you have to
know the referent's predicate to know the available properties on a
statement).