Export package utilities
The package provides lots of utilities to start building packages.
This shifts the main focus from the browser application into the package
SDK. This package provides a lot of the new functions used by the
link-solid-boilerplate and is the new place for more experimental
features which should be updates across all downstream projects but
aren't a good fit for link-lib/link-redux.